change coins by flip the card

+++ plugins/simple_networking/lib/modules/remote_config/models/rewards_asset_model.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'rewards_asset_model.freezed.dart';
@@ -7,7 +8,7 @@ part 'rewards_asset_model.g.dart';
 class RewardsAssetModel with _$RewardsAssetModel {
   factory RewardsAssetModel({
     required String name,
-    required double value,
+    required Decimal value,
   }) = _RewardsAssetModel;
 
   factory RewardsAssetModel.fromJson(Map<String, dynamic> json) => _$RewardsAssetModelFromJson(json);
